Notes in description
I added hiding password and at the end of drivers login is called (before if you signed out from another user and then went into driver you go back to the previous user) i also removed exiting when logging in as i dont think it is necessary and i also changed the format to be consistent (mostly). I also added some comments
In dispatchers the user is just asked to enter service number but i dont think this is correct as service number is not unique - the primary key is a tuple. I believe we need to list out the services along with the driver/driver id and the user can select a service by entering a number